requirements.txt
python-telegram-bot==13.7
beautifulsoup4==4.12.0
requests==2.31.0
pandas==2.0.3
schedule==1.2.0
python-dotenv==1.0.0